Create SlaveArduino.ino
This is a simple code I have created to send data from one Arduino Uno to another Arduino Uno using the UART protocol. I will use this as a basis for communication between the flight controller and the Arduino Uno. I will build on this code to have it read data from the GPS and pass it to the flight controller.