Create MasterArduino.ino
This is a simple code I have created to receive data from one Arduino Uno to another Arduino Uno using the UART protocol. This code will only be used to show that the slave Arduino Uno is transmitting the data appropriately. The Arduino Uno using this code will eventually be replaced with the flight controller of the drone.