Update MasterArduino.ino
The sole purpose of this program is to verify that the Arduino Uno that is used as a slave is able to obtain data from the GPS and pass the live coordinates to another Arduino Uno through the UART protocol. Only update done was extending the length of the data received from the Arduino Uno slave. This program will no longer be worked on.