Create LedGpsLocation
This project lights up LEDs when they are within certain GPS coordinates. I have chosen a square shape so four LEDs will be utilized for this program. The GPS reads coordinates up to the sixth decimal value, I have modified the program so that the only values we are interested are up to the fifth decimal value as to allow the GPS flexibility when reaching said coordinates. The values of the GPS have been replaced with # for safety reasons. Additional modification to this program will be to assimilate it with that of SlaveArduino.ino (https://github.com/David-Ceniceros/autonomous-flight-trajectory-using-GPS/blob/main/SlaveArduino.ino).