Create Desired_positioning.cpp
This program is to be used on the Cygnus IDE. This code will eventually be used to allow the drone to autonomously maneuver the PlutoX drone. The shape I have selected is a square. The drone will acknowledge when it has reached certain coordinates and will move to the next coordinates. This is currently a work in progress as I will need to incorporate the GPS to the PlutoX flight controller.