Create light_External_LED.cpp
This allows for an external LED connected to the flight controller to light up. LEDs are attached to the flight controller board. This project was created to test the connections and abilities of the PlutoX flight controller. Future improvements of this project will be to test the UART pin and being able to have an external LED light up when the flight controller receives a signal from an Arduino Uno.