Update and rename Example-TinyGPS.ino to lat_lng_to_UTM
This code was constructed by integrating the TinyGPS example for Arduino(https://github.com/sparkfun/SparkFun_I2C_GPS_Arduino_Library/blob/master/examples/Example2-TinyGPS/Example2-TinyGPS.ino) and the UTM to WGS84 conversion (https://alephnull.net/software/gis/UTM_WGS84_C_plus_plus.shtml). I have removed the feature which indicated the time and date since it is not needed for this project. This program utilizes libraries to obtain data from the GPS and equations to convert that information into UTM coordinates which are displayed in Kilometers. Both GPS and UTM coordinates are displayed instantaneously and are displayed every second. If GPS coordinates are not obtained, a message will appear in the monitor informing the user that the coordinates are not obtained. There will be no more modifications done to this program unless they are to reorganize the program.