Update Desired_positioning.cpp
This program is to be used on the Cygnus IDE. This code will display messages on the monitor to demonstrate the autonomously trajectory of the plutoX drone. The trajectory will be in a shape of a square. A timer will indicate when the drone is to move to a location. This is currently a work in progress as I will need to incorporate the GPS to the plutoX flight controller and will replace the messages with commands.