Update and rename Desired_positioning.cpp to trajectory_subproject.cpp
This program is to be used on the Cygnus IDE. This program will allow the drone to autonomously fly in one direction for 3 seconds, stop for 1.5 seconds, return to its starting position, then have the drone land. Serial monitor messages have been commented out which would describe the actions of the drone within their appropriate time. The trajectory of this program is a straight line, but updates will be performed to have it flying in a square. This program utilizes a timer that will indicate when the drone is to move to a location.