Update trajectory_subproject.cpp
This program is to be used on the Cygnus IDE. This program will allow the drone to autonomously fly in a shape of a square. The drone will fly for 3 seconds in one direction, stop for 1.5 seconds, then continue to its next position until a the trajectory in the shape of a square has been completed. After it has completed the trajectory, the drone is programmed to land. Serial monitor messages have been commented out which would describe the actions of the drone within their appropriate time. No further updates will be done to this program due to the allotted time of the project and software complications that arose.