Create SlaveArduino_simple.ino
This program was used to send a signal to another Arduino Uno board. The other Arduino Uno board has an LED light attached to it which would light up whenever the signal was received. This program is to replace (https://github.com/David-Ceniceros/autonomous-flight-trajectory-using-GPS/blob/main/SlaveArduino.ino) which was able to send the coordinates to the other Arduino Uno board but not to the PlutoX flight controller. Further progress is to be made to complete that project.